pyproject.toml
[project]
name = "connect-four-agent"
version = "0.1.0"
description = "Connect four arena for LLM agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "gradio>=5.49.1",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.4.10",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "urllib3>=2.5.0",
]
